+++ Tests/xcodeprojTests/Objects/SwiftPackage/XCRemoteSwiftPackageReferenceTests.swift
@@ -4,12 +4,32 @@ import XCTest
 @testable import XcodeProj
 
 final class XCRemoteSwiftPackageReferenceTests: XCTestCase {
+    func test_init() throws {
+        // Given
+        let decoder = XcodeprojPropertyListDecoder()
+        let plist: [String: Any] = ["reference": "ref",
+                                    "repositoryURL": "url",
+                                    "requirement": [
+                                        "kind": "revision",
+                                        "revision": "abc",
+                                    ]]
+        let data = try PropertyListSerialization.data(fromPropertyList: plist, format: .xml, options: 0)
+
+        // When
+        let got = try decoder.decode(XCRemoteSwiftPackageReference.self, from: data)
+
+        // Then
+        XCTAssertEqual(got.reference.value, "ref")
+        XCTAssertEqual(got.repositoryURL, "url")
+        XCTAssertEqual(got.versionRules, XCRemoteSwiftPackageReference.VersionRules.revision("abc"))
+    }
+
     func test_versionRules_returnsTheRightPlistValues_when_revision() throws {
         // When
         let subject = XCRemoteSwiftPackageReference.VersionRules.revision("sha")
 
         // Given
-        let got = try subject.plistValues()
+        let got = subject.plistValues()
 
         // Then
         XCTAssertEqual(got, [
@@ -23,7 +43,7 @@ final class XCRemoteSwiftPackageReferenceTests: XCTestCase {
         let subject = XCRemoteSwiftPackageReference.VersionRules.branch("master")
 
         // Given
-        let got = try subject.plistValues()
+        let got = subject.plistValues()
 
         // Then
         XCTAssertEqual(got, [
@@ -37,7 +57,7 @@ final class XCRemoteSwiftPackageReferenceTests: XCTestCase {
         let subject = XCRemoteSwiftPackageReference.VersionRules.exact("3.2.1")
 
         // Given
-        let got = try subject.plistValues()
+        let got = subject.plistValues()
 
         // Then
         XCTAssertEqual(got, [
@@ -51,7 +71,7 @@ final class XCRemoteSwiftPackageReferenceTests: XCTestCase {
         let subject = XCRemoteSwiftPackageReference.VersionRules.upToNextMajorVersion("3.2.1")
 
         // Given
-        let got = try subject.plistValues()
+        let got = subject.plistValues()
 
         // Then
         XCTAssertEqual(got, [
@@ -65,7 +85,7 @@ final class XCRemoteSwiftPackageReferenceTests: XCTestCase {
         let subject = XCRemoteSwiftPackageReference.VersionRules.range(from: "3.2.1", to: "4.0.0")
 
         // Given
-        let got = try subject.plistValues()
+        let got = subject.plistValues()
 
         // Then
         XCTAssertEqual(got, [
@@ -80,12 +100,50 @@ final class XCRemoteSwiftPackageReferenceTests: XCTestCase {
         let subject = XCRemoteSwiftPackageReference.VersionRules.upToNextMinorVersion("3.2.1")
 
         // Given
-        let got = try subject.plistValues()
+        let got = subject.plistValues()
 
         // Then
         XCTAssertEqual(got, [
             "kind": "upToNextMinorVersion",
             "minimumVersion": .string(.init("3.2.1")),
         ])
+    }
+
+    func test_plistValues() throws {
+        // When
+        let proj = PBXProj()
+        let subject = XCRemoteSwiftPackageReference(repositoryURL: "repository",
+                                                    versionRules: .exact("1.2.3"))
+
+        // Given
+        let got = try subject.plistValues(proj: proj, reference: "ref")
+
+        // Then
+        XCTAssertEqual(got, [
+            "repositoryURL": "repository",
+            "requirement": .dictionary([
+                "kind": "exactVersion",
+                "minimumVersion": "1.2.3",
+            ]),
+        ])
+    }
+
+    func test_equal() {
+        // When
+        let first = XCRemoteSwiftPackageReference(repositoryURL: "repository",
+                                                  versionRules: .exact("1.2.3"))
+        let second = XCRemoteSwiftPackageReference(repositoryURL: "repository",
+                                                   versionRules: .exact("1.2.3"))
+
+        // Then
+        XCTAssertEqual(first, second)
+    }
+
+    func test_name() {
+        // When
+        let subject = XCRemoteSwiftPackageReference(repositoryURL: "https://github.com/tuist/xcodeproj", versionRules: nil)
+
+        // Then
+        XCTAssertEqual(subject.name, "xcodeproj")
     }
 }