+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `override` flag to `PBXGroup.addFile(at:,sourceTree:,sourceRoot:)` https://github.com/tuist/xcodeproj/pull/410 by @mrylmz
 - Rename some internal variables to have a more representative name https://github.com/tuist/xcodeproj/pull/415 by @pepibumur.
 - **Breaking** Change the UUID generation logic to generate ids with a length of 24 https://github.com/tuist/xcodeproj/pull/432 by @pepibumur.
+
 ### Added
 
 - **Breaking** Swift 5 support https://github.com/tuist/xcodeproj/pull/397 by @pepibumur.
@@ -21,6 +22,8 @@
 - **Breaking** Added throwing an error in case group path can't be resolved by @damirdavletov  
 - **Breaking** Added remote project support to PBXContainerItemProxy by @damirdavletov
 - **Breaking** Add support for `RemoteRunnable` https://github.com/tuist/xcodeproj/pull/400 by @pepibumur.
+- Support for Swift Package Manager https://github.com/tuist/xcodeproj/pull/439 by @pepibumur.
+
 ### Fixed
 
 - Carthage integration https://github.com/tuist/xcodeproj/pull/416 by @pepibumur.